Move corpus encode/decode to importable fuzz/corpus package
Move encode/decode logic from package main (corpus.go) into
fuzz/corpus/corpus.go as an importable package. Both fuzz.go and
cmd/seedcorpus now import corpus.Encode/corpus.Decode, eliminating
duplication. Update gitignores to allow Go files in corpus/ while
ignoring generated .jfr files.
